Moved `Tup` and its axioms to own module
This prevents theorems in the `fun` module from accessing the inner values by accident.
